Add correct homework for task2
