+++ Itacademycourses/src/by/it/academy/hw2/Task5.java
@@ -1,0 +1,14 @@
+package by.it.academy.hw2;
+
+import java.util.Scanner;
+
+public class Task5 {
+    public static void main(String[] args) {
+
+        Scanner in = new Scanner(System.in);
+
+        int num = in.nextInt();
+
+        System.out.println(num % 2);
+    }
+}
